refactor: move POI helpers to core/world_runtime_poi.js and import them; remove inline implementations
